Moving interfaces into domain model
