+++ app/src/main/java/Domain/Data/ExerciseMeasure.java
@@ -1,4 +1,4 @@
-package Domain;
+package Domain.Data;
 
 public class ExerciseMeasure {
