+++ app/src/main/java/Domain/Data/FitnessTest.java
@@ -1,4 +1,4 @@
-package Domain;
+package Domain.Data;
 
 import java.util.Date;
